+++ applovin_max/android/src/main/java/com/applovin/applovin_max/AppLovinMAX.java
@@ -1406,9 +1406,8 @@ public class AppLovinMAX
         MaxError error = response.getError();
         if ( error != null )
         {
-            Map<String, Object> errorObject = new HashMap<>( 3 );
+            Map<String, Object> errorObject = new HashMap<>( 2 );
             errorObject.put( "message", error.getMessage() );
-            errorObject.put( "adLoadFailureInfo", error.getAdLoadFailureInfo() );
             errorObject.put( "code", error.getCode() );
 
             networkResponseObject.put( "error", errorObject );